+++ .dev/readme.mjs
@@ -8,16 +8,67 @@ import { readme } from './templates/readme.mjs'
 import {
 	pkg, 
 	addTextBetweenAMark,
-	execProcess, 
+	execProcess,
+	readFile,
+	existPath,
+	joinPath,
+	paths,
+	joinUrl, 
 } from '@backan/config/core'
 
 await execProcess( {
 	name : 'CHANGE README',
 	on   : async ( ) => {
 
-		const readmeTemp    = readme( pkg )
-		const convertReadme = async filePath => {
+		const getContent = content => {
 
+			const lines = content.split( '\n' )
+				.filter( line => !line.startsWith( ':::' ) ) // Filtrar líneas que no empiezan por ':::'
+			const index = lines.findIndex( line => line.startsWith( '#' ) )
+			return index === -1 ? content : lines.slice( index + 1 ).join( '\n' ).trim()
+		
+		}
+
+		const replaceRelativeUrls = ( inputString, baseUrl, basePath ) =>{
+
+			const regex = /(src="|]\()(\.{1,2}|\/)([^\s'")]+)/g
+
+			const result = inputString.replace( regex, ( match, prefix, relativeType, path ) => {
+
+				path = path.replace( /\.md$/, '' )
+				// console.log( {
+				// 	prefix,baseUrl, path, relativeType,
+				// } )
+				if ( relativeType === '/' ) return prefix + '' + joinUrl( baseUrl, path )
+
+				if ( relativeType === '..' ) return prefix + '' + joinUrl( baseUrl, basePath.split( '/' )[0] ,path ) 
+	
+				return prefix + '' + joinUrl( baseUrl, basePath, path ) 
+			
+			} )
+		
+			return result
+		
+		}
+		const readmeTemp = readme( pkg )
+
+		const convertReadme = async id => {
+			
+			const filePath   = id === 'monorepo' ? 'README.md' : joinPath( paths.workspaceDir,'packages', id, 'README.md' )
+			const docsPathID = id === 'backan' ? 'core' : id
+			const docsPath   = joinPath( paths.workspaceDir, 'docs', 'guide', docsPathID, 'index.md' ) 
+			const existsDocs = await existPath( docsPath )
+
+			if( existsDocs ){
+
+				const content = replaceRelativeUrls( 
+					getContent( await readFile( docsPath ) ), 
+					pkg.data.homepage, 
+					joinPath( 'guide', docsPathID ), 
+				)
+				await addTextBetweenAMark( filePath, '<!-- PIGEONPOSSE START DOCS -->', '<!-- PIGEONPOSSE END DOCS -->', content )
+			
+			}
 			await addTextBetweenAMark( filePath, '<!-- PIGEONPOSSE START MARK -->', '<!-- PIGEONPOSSE END MARK -->', readmeTemp.mark )
 			await addTextBetweenAMark( filePath, '<!-- PIGEONPOSSE START CONTENT -->', '<!-- PIGEONPOSSE END CONTENT -->', readmeTemp.content )
 			await addTextBetweenAMark( filePath, '<!-- PIGEONPOSSE START INDEX -->', '<!-- PIGEONPOSSE END INDEX -->', readmeTemp.index )
@@ -26,20 +77,20 @@ await execProcess( {
 		
 		}
 
-		const readmePaths = [
-			'README.md',
-			'packages/_config/README.md',
-			'packages/backan/README.md',
-			'packages/builder/README.md',
-			'packages/core/README.md',
-			'packages/create/README.md',
-			'packages/docs/README.md',
-			'packages/server/README.md',
+		const ids = [
+			'monorepo',
+			'_config',
+			'backan',
+			'builder',
+			'core',
+			'create',
+			'docs',
+			'server',
 		]
 
-		for ( const path of readmePaths ) {
+		for ( const id of ids ) {
 
-			await convertReadme( path )
+			await convertReadme( id )
  
 		}