+++ packages/builder/examples/app.ts
@@ -1,8 +1,12 @@
-import { App } from '@backan/core'
+import { App }         from '@backan/core'
+import { route }       from './route'
+import type { AppEnv } from './env'
 
-type AppEnv = {DB: object}
-export const app = new App<AppEnv>( {
+const app = new App<AppEnv>( {
 	version     : '1.0.0',
 	title       : 'BACKAN Example app',
 	description : 'API documentation for BACKAN Example',
 } )
+app.addRoute( route )
+
+export { app }