+++ packages/builder/src/main.ts
@@ -35,6 +35,7 @@ import { BuildError }          from './error'
 
 export { buildSchema } from './schema'
 export { buildMD } from './md'
+export { createClient } from './client'
 export type * from './types'
 
 /**